+++ classes/Generator/Task/Generate/Unittest.php
@@ -21,6 +21,14 @@ class Generator_Task_Generate_Unittest extends Task_Generate
 	);
 
 	/**
+	 * @var  array  Arguments mapped to options
+	 */
+	protected $_arguments = array(
+		1 => 'name',
+		2 => 'groups',
+	);
+
+	/**
 	 * Validates the task options.
 	 *
 	 * @param   Validation  $validation  The validation object to add rules to