+++ classes/Task/Generate/Controller.php
@@ -3,36 +3,36 @@
  * Generates application controllers from templates. The controller can be
  * created in either the application folder or a module folder.
  *
- * Additional options:
+ * <comment>Additional options:</comment>
  *
- *   --name=CONTROLLER (required)
+ *   <info>--name=CONTROLLER</info> <alert>(required)</alert>
  *
  *     The name of this controller. If 'Controller_' is not included in
  *     the name, it will be prepended automatically.
  *
- *   --extend=CLASS
+ *   <info>--extend=CLASS</info>
  *
  *     The name of the parent class from which this is extended, if none
  *     is given then Controller will be used by default.
  *
- *   --actions=ACTION[,ACTION[,...]]
+ *   <info>--actions=ACTION[,ACTION[,...]]</info>
  *
  *     A comma-separated list of optional action methods to be included in
  *     this controller, without the 'action_' prefix.
  * 
- *   --blank
+ *   <info>--blank</info>
  *
  *     The skelton methods will be omitted if this option is set.
  * 
- * Examples
+ * <comment>Examples</comment>
  * ========
- * minion generate:controller --name=Home
+ * <info>minion generate:controller --name=Home</info>
  *
  *     class : Controller_Home extends Controller
  *     file  : APPPATH/classes/Controller/Home.php
  *
- * minion generate:controller --name=Home --module=logger --blank \
- *     --extend=Controller_Template --actions="index,create,edit"
+ * <info>minion generate:controller --name=Home --module=logger --blank \
+ *     --extend=Controller_Template --actions="index, create, edit"</info>
  *
  *     class : Controller_Home extends Controller_Template
  *     file  : MODPATH/logger/classes/Controller/Home.php