+++ classes/Task/Generate/Controller/View.php
@@ -3,21 +3,21 @@
  * Generates template controllers with associated view files. The files can
  * be created in either the application folder or a module folder.
  *
- * Additional options:
+ * <comment>Additional options:</comment>
  *
- *   --name=CONTROLLER (required)
+ *   <info>--name=CONTROLLER</info> <alert>(required)</alert>
  *
  *     The name of this controller. If 'Controller_' is not included in
  *     the name, it will be prepended automatically.
  *
- *   --actions=ACTION[,ACTION[,...]]
+ *   <info>--actions=ACTION[,ACTION[,...]]</info>
  *
  *     A comma-separated list of optional action methods to be included in
  *     this controller, without the 'action_' prefix.
  * 
- * Examples
+ * <comment>Examples</comment>
  * ========
- * minion generate:controller:view --name=Home --actions="index,create,edit"
+ * <info>minion generate:controller:view --name=Home --actions="index, create, edit"</info>
  *
  *     class : Controller_Home extends Controller_Template
  *     file  : APPPATH/classes/Controller/Home.php