+++ classes/Task/Generate/Model.php
@@ -4,38 +4,38 @@
  * either the application folder or a module folder, and can optionally be
  * configured for transparent extension.
  *
- * Additional options:
+ * <comment>Additional options:</comment>
  *
- *   --name=MODEL (required)
+ *   <info>--name=MODEL</info> <alert>(required)</alert>
  *
  *     The name of this Model. If 'Model' is not included in the name, it
  *     will be prepended automatically.
  *
- *   --extend=CLASS
+ *   <info>--extend=CLASS</info>
  *
  *     The name of the parent class from which this is optionally extended.
  *
- *   --stub=MODEL
+ *   <info>--stub=MODEL</info>
  *
  *     If set, this stub will be created as a transparent extension of the 
  *     model; the 'Model' prefix may also be omitted.
  *
- *   --no-test
+ *   <info>--no-test</info>
  *
  *     A test case will be created automatically for the class unless this
  *     option is set.
  *
- * Examples
+ * <comment>Examples</comment>
  * ========
- * minion generate:model --name=Log
+ * <info>minion generate:model --name=Log</info>
  *
  *     class : Model_Log extends Model
  *     file  : APPPATH/classes/Model/Log.php
  *     class : Model_LogTest extends Unittest_TestCase
  *     file  : APPPATH/tests/Model/LogTest.php 
  *
- * minion generate:model --name=Logger_Model_Log --extend=Model_Database \
- *     --module=logger --stub=Model_Log --no-test
+ * <info>minion generate:model --name=Logger_Model_Log --extend=Model_Database \
+ *     --module=logger --stub=Model_Log --no-test</info>
  *
  *     class : Logger_Model_Log extends Model_Database
  *     file  : MODPATH/logger/classes/Logger/Model/Log.php