+++ views/generator/type_class.php
@@ -10,8 +10,8 @@
 <?php 
 	echo ( ! empty($abstract) ? 'abstract class ' : 'class ');
 	echo $name;	
-	if ( ! empty($extends)) {echo ' extends '.$extends;} 
-	if ( ! empty($implements)) {echo ' implements '.$implements;}
+	if ( ! empty($extends)) {echo ' extends ', $extends;} 
+	if ( ! empty($implements)) {echo ' implements ', $implements;}
 	if ( ! empty($blank)) {echo ' {}';} else { ?> 
 {
 	/**
@@ -36,7 +36,7 @@
 <?php	if (isset($methods[$group])) foreach ($methods[$group] as $method => $m): ?>
 	<?php echo $m['doccomment'] ?> 
 	<?php echo $m['signature'] ?>
-<?php if ($m['abstract']): echo ';'.PHP_EOL.PHP_EOL; else: ?> 
+<?php if ($m['abstract']): echo ';', PHP_EOL, PHP_EOL; else: ?> 
 	{
 		<?php echo $m['body'] ?> 
 	}